Finish component separation.
- Move non-interface components out of the main source folder.
- Create project library.
- Create dependency hierarchy between components.
- Don't include all ICU libraries.
