+++ src/data/table.cpp
@@ -1,5 +1,5 @@
 #include "table.h"
-#include "util.h"
+#include "mapper.h"
 #include <iomanip>
 #include <algorithm>
 #include <sstream>
