+++ src/font/CMakeLists.txt
@@ -11,10 +11,11 @@ file(GLOB SABLE_FONT_SOURCE_FILES
     normalize.h
     error.cpp
     error.h
+    codenotfound.h
 
 )
 
 add_library(sable_font STATIC ${SABLE_FONT_SOURCE_FILES})
 
-target_link_libraries(sable_font PUBLIC ${ICU_LIBRARIES})
+target_link_libraries(sable_font PUBLIC ${SABLE_ICU_DEPS})
 target_include_directories(sable_font PUBLIC ${SABLE_INCLUDE_DIR} ${YAML_INCLUDE_DIR} ${ICU_INCLUDE_DIRS})
