+++ src/font/builder.cpp
@@ -1,6 +1,6 @@
 #include "builder.h"
-#include "exceptions.h"
 #include "normalize.h"
+#include "codenotfound.h"
 
 namespace sable {
 
@@ -13,14 +13,45 @@ FontError generateError(YAML::Mark m, Args ...args)
 }
 
 
+enum class SubfieldErrorType {
+    Integer,
+    IntegerSequence,
+    FollowConstraint
+};
+std::string getErrorTypeLabel(SubfieldErrorType type)
+{
+    using Type = SubfieldErrorType;
+    switch (type) {
+    case Type::Integer:
+        return "an integer.";
+    case Type::IntegerSequence:
+        return "a sequence of integers.";
+    case Type::FollowConstraint:
+        return "in alignment with a constraint.";
+    default:
+        throw std::logic_error("Undefined option.");
+    }
+}
+
+
 template<class T>
 struct SubfieldError : std::runtime_error {
     std::string fieldName;
     std::string parentName;
-    std::string type;
-    SubfieldError(std::string fieldName, std::string type): std::runtime_error(fieldName + " must be a " + type) {
+    using Type = SubfieldErrorType;
+    Type type;
+    std::string constraint;
+    SubfieldError(std::string fieldName, Type type):
+        constraint{getErrorTypeLabel(type)},
+        std::runtime_error(fieldName + " must be " + getErrorTypeLabel(type)) {
         this->fieldName = fieldName;
         this->type = type;
+    }
+    SubfieldError(std::string fieldName, std::string constraint):
+        type{Type::FollowConstraint},
+        std::runtime_error(fieldName + ' ' + constraint) {
+        this->fieldName = fieldName;
+        this->constraint = constraint;
     }
 };
 
@@ -60,15 +91,11 @@ void generate(
         } else {
             for (auto it = node.begin(); it != node.end(); ++it) {
                 std::string str = it->first.as<std::string>();
-                if (str == "length") {
-                    int test = 1;
-                }
+
                 try {
                     adder(str, it->second.as<T>());
                 } catch (YAML::TypedBadConversion<T>) {
                     throw generateError(node.Mark(), name, field, str, msg);
-                } catch (ConfigError &e) {
-                    throw generateError(node.Mark(), name, "", e.what());
                 } catch (sable::SubfieldError<T> &e) {
                     e.parentName = str;
                     throw e;
@@ -91,7 +118,7 @@ Font::Page buildPage(
 
     auto encodingConv = [&page, commandValue] (const std::string& id, Font::TextNode&& n) {
         if (n.code == commandValue) {
-            throw ConfigError("glyphs cannot have a code that matches the command value.");
+            throw SubfieldError<Font::TextNode>(Font::CODE_VAL, "matches the command value, which is not allowed.");
         }
         std::string newId = id;
         if (newId.front() == '[') {
@@ -111,12 +138,21 @@ Font::Page buildPage(
             // but better safe than sorry
             throw generateError(e.mark, name, e.field, e.type);
         } catch (sable::SubfieldError<Font::TextNode>& e) {
+            if (e.type == SubfieldErrorType::FollowConstraint) {
+                throw generateError(
+                    node.Mark(),
+                    name,
+                    Font::ENCODING,
+                    e.parentName,
+                    std::string("") + "has a \"" + e.fieldName + "\" field that " + e.constraint
+                );
+            }
             throw generateError(
                 node.Mark(),
                 name,
                 Font::ENCODING,
                 e.parentName,
-                std::string("") + "has a \"" + e.fieldName + "\" field that is not "  + e.type
+                std::string("") + "has a \"" + e.fieldName + "\" field that is not "  + e.constraint
             );
         }
 
@@ -248,12 +284,21 @@ Font Builder::make(const YAML::Node &config, const std::string &name, const std:
             try {
                 f.addPage(buildPage(name, page, commandCode, byteWidth));
             } catch (sable::SubfieldError<Font::TextNode>& e) {
+                if (e.type == SubfieldErrorType::FollowConstraint) {
+                    throw generateError(
+                        page.Mark(),
+                        name,
+                        std::string{Font::PAGES} + " #" + std::to_string(pageIdx),
+                        e.parentName,
+                        std::string("") + "has a \"" + e.fieldName + "\" field that is not "  + e.constraint
+                    );
+                }
                 throw generateError(
                     page.Mark(),
                     name,
                     std::string{Font::PAGES} + " #" + std::to_string(pageIdx),
                     e.parentName,
-                    std::string("") + "has a \"" + e.fieldName + "\" field that is not "  + e.type
+                    std::string("") + "has a \"" + e.fieldName + "\" field that is not "  + e.constraint
                 );
             }
             ++pageIdx;
@@ -308,13 +353,13 @@ namespace YAML {
             try {
                 rhs.code = node[CODE_VAL].as<unsigned int>();
             } catch(YAML::TypedBadConversion<unsigned int> &e) {
-                throw sable::SubfieldError<Font::TextNode>(CODE_VAL, "an integer.");
+                throw sable::SubfieldError<Font::TextNode>(CODE_VAL, sable::SubfieldErrorType::Integer);
             }
             if (node[TEXT_LENGTH_VAL].IsDefined()) {
                 try {
                     rhs.width = node[TEXT_LENGTH_VAL].as<int>();
                 } catch(YAML::TypedBadConversion<int> &e) {
-                    throw sable::SubfieldError<Font::TextNode>(TEXT_LENGTH_VAL, "an integer.");
+                    throw sable::SubfieldError<Font::TextNode>(TEXT_LENGTH_VAL, sable::SubfieldErrorType::Integer);
                 }
             }
             return true;
@@ -322,7 +367,7 @@ namespace YAML {
             try {
                 rhs.code = node.as<unsigned int>();
             } catch(YAML::TypedBadConversion<unsigned int> &e) {
-                throw sable::SubfieldError<Font::TextNode>(CODE_VAL, "an integer.");
+                throw sable::SubfieldError<Font::TextNode>(CODE_VAL, sable::SubfieldErrorType::Integer);
             }
             return true;
         }
@@ -371,18 +416,18 @@ namespace YAML {
             try {
                rhs.width = node[TEXT_LENGTH_VAL].as<int>();
             } catch (YAML::TypedBadConversion<int> &e) {
-                throw sable::SubfieldError<sable::Font::NounNode>{TEXT_LENGTH_VAL, "an integer."};
+                throw sable::SubfieldError<sable::Font::NounNode>{TEXT_LENGTH_VAL, sable::SubfieldErrorType::Integer};
             }
         }
         if (!node[CODE_VAL].IsDefined() || !node[CODE_VAL].IsSequence()) {
-            throw sable::SubfieldError<sable::Font::NounNode>{CODE_VAL, "a sequence of integers."};
+            throw sable::SubfieldError<sable::Font::NounNode>{CODE_VAL, sable::SubfieldErrorType::IntegerSequence};
         }
         rhs.codes.reserve(node[CODE_VAL].size());
         for (auto it2 = node[CODE_VAL].begin(); it2 != node[CODE_VAL].end(); ++it2) {
             try {
                 rhs.codes.push_back(it2->as<int>());
             } catch (YAML::TypedBadConversion<int> &e) {
-                throw sable::SubfieldError<sable::Font::NounNode>{CODE_VAL, "a sequence of integers."};
+                throw sable::SubfieldError<sable::Font::NounNode>{CODE_VAL, sable::SubfieldErrorType::IntegerSequence};
             }
         }
         return true;
