+++ src/parse/CMakeLists.txt
@@ -8,8 +8,9 @@ file(GLOB SABLE_PARSING_SOURCE_FILES
     block.h
     block.cpp
     result.h
+    errorhandling.h
 )
 
 add_library(sable_parsing STATIC ${SABLE_PARSING_SOURCE_FILES})
-target_link_libraries(sable_parsing ${ICU_LIBRARIES})
+target_link_libraries(sable_parsing sable_data sable_font ${SABLE_ICU_DEPS})
 target_include_directories(sable_parsing PUBLIC ${SABLE_INCLUDE_DIR} ${ICU_INCLUDE_DIRS})
