+++ src/project/folder.h
@@ -5,7 +5,7 @@
 
 #include "data/table.h"
 #include "project/group.h"
-#include "util.h"
+#include "data/mapper.h"
 #include "wrapper/filesystem.h"
