+++ tests/CMakeLists.txt
@@ -8,6 +8,7 @@ file(GLOB SABLE_TEST_FILES
     catch/data/collisions.cpp
     catch/data/table.cpp
     catch/data/addresslist.cpp
+    catch/data/mapper.cpp
 
     catch/font/fonts.cpp
     catch/font/characteriterator.cpp
@@ -29,9 +30,8 @@ file(GLOB SABLE_TEST_FILES
     catch/project/handler.cpp
     catch/project/mapperconv.cpp
     catch/project/roms.cpp
-
-    catch/project.cpp
-    catch/mapping.cpp
+    catch/project/project.cpp
+    catch/project/util.cpp
 )
 
 include_directories(helpers)
@@ -39,8 +39,8 @@ include_directories(helpers)
 message(STATUS ${Catch2_INCLUDE_DIRS})
 
 add_executable(tests ${SABLE_TEST_FILES})
-target_compile_definitions(sable_lib PUBLIC USER_SABLE_TEST_HELPERS)
-target_link_libraries(tests sable_lib Catch2::Catch2)
+target_compile_definitions(sable_project PUBLIC USER_SABLE_TEST_HELPERS)
+target_link_libraries(tests sable_project Catch2::Catch2)
 target_include_directories(tests PUBLIC ${Catch2_INCLUDE_DIRS})
 
 configure_file("${CMAKE_CURRENT_SOURCE_DIR}/sample/sample.sfc" "sample.sfc" COPYONLY)
